Basic support for pa_ciml (C-in-ML).
Pa_ciml is a "C-in-ML" language_extension that makes it easier to
write FFI code. It is a syntax extension, but also generates
additional C files that must be included when building the project.
See [http://friggeri.net/blog/2008/02/26/ciml-c-in-ocaml].

If the `pa_ciml` language is specified (* LANG ... pa_ciml ... *),
then the script will include the generated `_ciml.c` file during the
linking stage.

Also it will look for an optional directive (* CIML ... *) and include
its value in the linkage call. This is primarily to include static
libraries, e.g. (* CIML -cclib -lsomelib *).